fix: replaces geom_path for geom_line in safety outlier explorer
According to ggplot doc: geom_path() connects the observations in the order in which they appear in the data. geom_line() connects them in order of the variable on the x axis. 

For this plot is more intuitive if lines are plotted with sorted by the x values and not to data appeareance.

Fix: #85